pyproject.toml
[tool.poetry]
name = "openai-examples"
version = "0.0.0"
description = "openai-examples"
authors = ["Harshad Hegde <hhegde@lbl.gov>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
setuptools = ">=65.5.0"
tox = "^3.25.1"
click = "^8.1.3"
importlib-metadata = "^4.8.0"
python-dotenv = "^1.0.0"
openai = "^0.27.6"
click-shell = "^2.1"
poetry = "^1.4.2"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
sphinx = {version = "^5.3.0", extras = ["docs"]}
sphinx-rtd-theme = {version = "^1.0.0", extras = ["docs"]}
sphinx-autodoc-typehints = {version = "^1.19.4", extras = ["docs"]}
sphinx-click = {version = "^4.3.0", extras = ["docs"]}
myst-parser = {version = "^0.18.1", extras = ["docs"]}

[tool.poetry.scripts]
chatbot = "openai-examples.cli:main"

[tool.poetry.extras]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
    "sphinx-click",
    "myst-parser"
    ]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = true
reverse_relative = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
